requirements.txt
# Smart Health Monitoring System
# SIS4415 Final Project

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# GraphQL with Strawberry
strawberry-graphql[fastapi]==0.216.1

# Async MongoDB Driver
motor==3.3.2
pymongo==4.6.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.7

# Environment Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client (for testing)
httpx==0.25.2

# Date and Time Utilities
python-dateutil==2.8.2

# Logging and Monitoring
python-json-logger==2.0.7

# For Azure Deployment
gunicorn==21.2.0
aiofiles==23.2.1

# WebSocket Support for GraphQL Subscriptions
websockets==12.0
